Major refactor done
