+++ .github/workflows/testing.yml
@@ -60,4 +60,4 @@ jobs:
         run: |
           cp tests/hevelius.yaml.github hevelius/hevelius.yaml
           set
-          PGPASSWORD=postgres JWT_SECRET_KEY=foo PYTHONPATH=.:./flask python -m pytest -s -v
+          PGPASSWORD=postgres JWT_SECRET_KEY=foo PYTHONPATH=. python -m pytest -s -v
