+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project versioning adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## 0.6.0 (unreleased)
+- Refactor: REST API moved to `hevelius.api` (run with `python -m hevelius.api`);
+  CLI handlers under `hevelius.cli`; shared domain modules (`passwords`, `equipment`,
+  `catalogs`, `asteroid`, `stats`). Removed top-level `heveliusbackend` package.
 - DB: schema bumped to 22: asteroids data now available.
 - DB: schema bumped to 23: camera rotation, camera details in the projects.
 - DB: schema bumped to 24: asteroid tags (`asteroid_tags`, `asteroid_tag_map`) for
