+++ doc/gunicorn-hevelius.service
@@ -10,7 +10,7 @@ Environment="PATH=/home/hevelius/devel/hevelius-backend:/home/hevelius/deve/heve
 ExecStart=/home/hevelius/devel/hevelius-backend/venv/bin/gunicorn \
   --keyfile /home/hevelius/tls/privkey.pem --certfile /home/hevelius/tls/cert.pem \
   --access-logfile /var/log/hevelius/access.log --error-logfile /var/log/hevelius/error.log \
-  --workers 1 --bind 0.0.0.0:5001 heveliusbackend.app:app
+  --workers 1 --bind 0.0.0.0:5001 hevelius.api:app
 ExecReload=/bin/kill -s HUP $MAINPID
 KillMode=mixed
 TimeoutStopSec=5
