+++ hevelius/api/__main__.py
@@ -1,0 +1,5 @@
+"""python -m hevelius.api"""
+from hevelius.api import app
+
+if __name__ == "__main__":
+    app.run()
