+++ hevelius/api/templates/histogram.html
