+++ hevelius/cli/asteroid.py
@@ -1,0 +1,2 @@
+"""CLI entry points for asteroid commands (logic lives in hevelius.asteroid)."""
+from hevelius.asteroid import *  # noqa: F401,F403
