+++ hevelius/cli/equipment.py
@@ -1,0 +1,2 @@
+"""CLI wrappers for equipment management (filters, sensors, telescopes, projects)."""
+from hevelius.equipment import *  # noqa: F401,F403
