+++ hevelius/cli/stats.py
@@ -1,0 +1,2 @@
+"""CLI entry points for stats/histogram commands (logic lives in hevelius.stats)."""
+from hevelius.stats import *  # noqa: F401,F403
