+++ tests/test_api.py
@@ -10,7 +10,7 @@ from datetime import datetime
 from flask_jwt_extended import create_access_token
 from tests.dbtest import use_repository
 from hevelius import db
-from heveliusbackend.app import app
+from hevelius.api import app
 
 
 class TestTaskAdd(unittest.TestCase):
